+++ src/components/index.js
@@ -1,3 +1,4 @@
+export { default as Asterisk } from './Asterisk';
 export { default as Loading } from './Loading';
 export { default as SortableList } from './SortableList';
 export { default as ReferredRecord } from './ReferredRecord/ReferredRecord';